1、代码临时提交

+++ src/main/java/cn/berry/rapids/eventbus/EventBusBuilder.java
@@ -24,7 +24,7 @@ public class EventBusBuilder {
         return this;
     }
 
-    public EventBusBuilder submitEventWaitTime(int val) {
+    public EventBusBuilder submitEventWaitTime(long val) {
         this.submitEventWaitTime = val;
         return this;
     }
